… doc/advopt.txt
… doc/docgen.md
… doc/markdown_rst.md
… lib/packages/docutils/dochelpers.nim
… lib/packages/docutils/rst.nim
… lib/packages/docutils/rstgen.nim
… lib/packages/docutils/rstidx.nim
… lib/pure/os.nim
… lib/std/private/osappdirs.nim
… lib/std/private/oscommon.nim
… lib/std/private/osdirs.nim
… lib/std/private/osfiles.nim
… lib/std/private/ospaths2.nim
… lib/std/private/osseps.nim
… lib/std/private/ossymlinks.nim
… lib/std/symlinks.nim
… nimdoc/extlinks/project/doc/manual.md
… nimdoc/extlinks/project/expected/_._/util.html
… nimdoc/extlinks/project/expected/_._/util.idx
… nimdoc/extlinks/project/expected/doc/manual.html
… nimdoc/extlinks/project/expected/doc/manual.idx
… nimdoc/extlinks/project/expected/main.html
… nimdoc/extlinks/project/expected/main.idx
… nimdoc/extlinks/project/expected/sub/submodule.html
… nimdoc/extlinks/project/expected/sub/submodule.idx
… nimdoc/extlinks/project/expected/theindex.html
… nimdoc/extlinks/project/main.nim
… nimdoc/extlinks/project/sub/submodule.nim
… nimdoc/extlinks/util.nim
… nimdoc/test_out_index_dot_html/expected/foo.idx